Used to split the master grave plot text file into individual grave files
The Grave marker list for Waldzell is in file- 
Friedhof Waldzell Grave List.txt

It is organized by the Friedhof plot, since the stone and family in plots will change over time.

It's format is still a work in progress. 
I find it much easier to maintain one large file instead of >100 small files, so
the large file is the original, and a pthon script is used to split theat file into 
plot files that are attached to RootsMagic sources. for each plot.

To run the python script- SplitMainGraveFileToIndividuals.py
make sure the output folder is created and empty, double click the py file.